requirements.txt
openai>=1.37.0
langchain==0.2.11
langchain-community==0.2.9
langchain-openai==0.1.21
pydantic>=2.7.0
python-dotenv>=1.0.0
chromadb>=0.5.3
tiktoken>=0.7.0
pdfminer.six>=20240706
rich>=13.7.1
orjson>=3.10.5
